requirement.txt
pathlib==1.0.1
schedule==1.1.0
dnspython==2.2.1
pymongo==4.2.0
python-dotenv==0.21.0
Pillow==7.1.2
torch==1.12.0
numpy==1.21.6
flask-cors==3.0.10
flask
waitress==2.1.2
torchvision==0.13.0
opencv-python==4.10.0.84
Werkzeug
